+++ internal/common/gamestate/gamestate.go
@@ -4,15 +4,8 @@ package gamestate
 import (
 	"github.com/SOMAS2020/SOMAS2020/internal/common/disasters"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/foraging"
-	"github.com/SOMAS2020/SOMAS2020/internal/common/roles"
 	"github.com/SOMAS2020/SOMAS2020/internal/common/shared"
 )
-
-type IIGOBaseRoles struct {
-	BasePresident roles.President
-	BaseSpeaker   roles.Speaker
-	BaseJudge     roles.Judge
-}
 
 // GameState represents the game's state.
 type GameState struct {
@@ -29,9 +22,6 @@ type GameState struct {
 	ClientInfos    map[shared.ClientID]ClientInfo
 	Environment    disasters.Environment
 	DeerPopulation foraging.DeerPopulationModel
-
-	// IIGOInfo contains states of the base versions of all roles
-	IIGOInfo IIGOBaseRoles
 
 	// [INFRA] add more details regarding state of game here
 	// REMEMBER TO EDIT `Copy` IF YOU ADD ANY REFERENCE TYPES (maps, slices, channels, functions etc.)
