+++ internal/server/iigointernal/legislature.go
@@ -10,8 +10,8 @@ import (
 	"github.com/pkg/errors"
 )
 
-+type legislative struct{
-	SpeakerId     shared.ClientID
+type legislature struct {
+	SpeakerID     shared.ClientID
 	budget        shared.Resources
 	judgeSalary   shared.Resources
 	RuleToVote    string
@@ -20,34 +20,34 @@ import (
 }
 
 // returnJudgeSalary returns the salary to the common pool.
-func (l *legislative) returnJudgeSalary() shared.Resources {
+func (l *legislature) returnJudgeSalary() shared.Resources {
 	x := l.judgeSalary
 	l.judgeSalary = 0
 	return x
 }
 
-func (l *legislative) withdrawJudgeSalary(gameState *gamestate.GameState) error {
+func (l *legislature) withdrawJudgeSalary(gameState *gamestate.GameState) error {
 	var judgeSalary = shared.Resources(rules.VariableMap["judgeSalary"].Values[0])
 	var withdrawError = WithdrawFromCommonPool(judgeSalary, gameState)
 	if withdrawError != nil {
-		featureSpeaker.judgeSalary = judgeSalary
+		l.judgeSalary = judgeSalary
 	}
 	return withdrawError
 }
 
-func (l *legislative) sendJudgeSalary() {
+func (l *legislature) sendJudgeSalary() {
 	amount, _ := l.clientSpeaker.PayJudge()
-	featureJudge.budget = amount
+	judicialBranch.budget = amount
 }
 
 // Receive a rule to call a vote on
-func (l *legislative) setRuleToVote(r string) {
+func (l *legislature) setRuleToVote(r string) {
 	l.RuleToVote = r
 }
 
 //Asks islands to vote on a rule
 //Called by orchestration
-func (l *legislative) setVotingResult(iigoClients map[shared.ClientID]baseclient.Client) {
+func (l *legislature) setVotingResult(iigoClients map[shared.ClientID]baseclient.Client) {
 
 	//TODO: Separate and clearly define speaker held information and vote held information (see ruleToVote)
 	//TODO: Remove tests
@@ -67,7 +67,7 @@ func (l *legislative) setVotingResult(iigoClients map[shared.ClientID]baseclient
 
 //Speaker declares a result of a vote (see spec to see conditions on what this means for a rule-abiding speaker)
 //Called by orchestration
-func (l *legislative) announceVotingResult() error {
+func (l *legislature) announceVotingResult() error {
 
 	rule, result, err := l.clientSpeaker.DecideAnnouncement(l.RuleToVote, l.VotingResult)
 
@@ -80,7 +80,7 @@ func (l *legislative) announceVotingResult() error {
 		l.VotingResult = false
 
 		//Perform announcement
-		broadcastToAllIslands(shared.TeamIDs[l.SpeakerId], generateVotingResultMessage(rule, result))
+		broadcastToAllIslands(shared.TeamIDs[l.SpeakerID], generateVotingResultMessage(rule, result))
 		return l.updateRules(rule, result)
 	}
 	return nil
@@ -89,7 +89,7 @@ func (l *legislative) announceVotingResult() error {
 //Example of the client implementation of DecideAnnouncement
 //A well behaved speaker announces what had been voted on and the corresponding result
 //Return "", _ for no announcement to occur
-func (l *legislative) DecideAnnouncement(ruleId string, result bool) (string, bool, error) {
+func (l *legislature) DecideAnnouncement(ruleId string, result bool) (string, bool, error) {
 	return ruleId, result, nil
 }
 
@@ -108,7 +108,7 @@ func generateVotingResultMessage(ruleID string, result bool) map[int]baseclient.
 	return returnMap
 }
 
-func (l *legislative) updateRules(ruleName string, ruleVotedIn bool) error {
+func (l *legislature) updateRules(ruleName string, ruleVotedIn bool) error {
 	l.budget -= 10
 	//TODO: might want to log the errors as normal messages rather than completely ignoring them? But then Speaker needs access to client's logger
 	notInRulesCache := errors.Errorf("Rule '%v' is not available in rules cache", ruleName)
@@ -134,7 +134,7 @@ func (l *legislative) updateRules(ruleName string, ruleVotedIn bool) error {
 
 }
 
-func (l *legislative) appointNextJudge(clientIDs []shared.ClientID) shared.ClientID {
+func (l *legislature) appointNextJudge(clientIDs []shared.ClientID) shared.ClientID {
 	l.budget -= 10
 	var election voting.Election
 	election.ProposeElection(baseclient.Judge, voting.Plurality)
