+++ internal/server/iigointernal/utilities.go
@@ -8,31 +8,6 @@ import (
 	"github.com/pkg/errors"
 )
 
-// PickUpRulesByVariable returns a list of rule_id's which are affected by certain variables
-func PickUpRulesByVariable(variableName string, ruleStore map[string]rules.RuleMatrix) ([]string, error) {
-	var Rules []string
-	if _, ok := rules.VariableMap[variableName]; ok {
-		for k, v := range ruleStore {
-			_, err := searchForStringInArray(variableName, v.RequiredVariables)
-			if err != nil {
-				Rules = append(Rules, k)
-			}
-		}
-		return Rules, nil
-	} else {
-		return []string{}, errors.Errorf("Variable name '%v' was not found in the variable cache", variableName)
-	}
-}
-
-func searchForStringInArray(val string, array []string) (int, error) {
-	for i, v := range array {
-		if v == val {
-			return i, nil
-		}
-	}
-	return 0, errors.Errorf("Not found")
-}
-
 func broadcastToAllIslands(sender shared.ClientID, data map[int]baseclient.Communication) {
 	islandsAlive := rules.VariableMap["islands_alive"]
 	for _, v := range islandsAlive.Values {
